requirements.txt
opencv-python~=4.7.0.72
numpy~=1.24.3
Pillow~=9.5.0
setuptools~=65.5.1